Updated dependencies and decorated logging messages

+++ lib/src/commands/clean.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:args/command_runner.dart';
+import 'package:cli_util/cli_logging.dart';
 
 import 'package:ngdart/util/ansipen.dart';
 
@@ -19,16 +20,23 @@ class CleanCommand extends Command<int> {
   FutureOr<int> run() async {
     var buildDir = Directory('build');
     var toolDir = Directory('.dart_tool');
+    var pubspec = File('pubspec.yaml');
+    var logger = Logger.standard(ansi: Ansi(true));
 
     try {
+      if (!(await pubspec.exists())) {
+        throw 'pubspec.yaml not found!';
+      }
       if ((await buildDir.exists())) {
-        print(progressLog + 'Deleting build/ ...');
+        var progress = logger.progress(progressLog + 'Deleting build/ ...');
         await buildDir.delete(recursive: true);
+        progress.finish(showTiming: true);
       }
       if ((await toolDir.exists()))
       {
-        print(progressLog + 'Deleting .dart_tools/ ...');
+        var progress = logger.progress(progressLog + 'Deleting .dart_tools/ ...');
         await toolDir.delete(recursive: true);
+        progress.finish(showTiming: true);
       }
     } catch (e) {
       throw Exception(e);
